--- examples/fl-tabular/centralized.py
@@ -1,119 +1,0 @@
-import pandas as pd
-import numpy as np
-from sklearn.model_selection import train_test_split
-from sklearn.preprocessing import LabelEncoder, StandardScaler
-import torch
-import torch.nn as nn
-import torch.optim as optim
-from torch.utils.data import DataLoader, TensorDataset
-
-url = "https://archive.ics.uci.edu/ml/machine-learning-databases/adult/adult.data"
-column_names = [
-    "age",
-    "workclass",
-    "fnlwgt",
-    "education",
-    "education_num",
-    "marital_status",
-    "occupation",
-    "relationship",
-    "race",
-    "sex",
-    "capital_gain",
-    "capital_loss",
-    "hours_per_week",
-    "native_country",
-    "income",
-]
-
-data = pd.read_csv(
-    url, header=None, names=column_names, na_values=" ?", sep=",\s", engine="python"
-)
-
-data.dropna(inplace=True)
-
-label_encoders = {}
-for column in data.select_dtypes(include=["object"]).columns:
-    label_encoders[column] = LabelEncoder()
-    data[column] = label_encoders[column].fit_transform(data[column])
-
-X = data.drop("income", axis=1).values
-y = data["income"].values
-
-X_train, X_test, y_train, y_test = train_test_split(
-    X, y, test_size=0.2, random_state=42
-)
-
-scaler = StandardScaler()
-X_train = scaler.fit_transform(X_train)
-X_test = scaler.transform(X_test)
-
-X_train_tensor = torch.tensor(X_train, dtype=torch.float32)
-X_test_tensor = torch.tensor(X_test, dtype=torch.float32)
-y_train_tensor = torch.tensor(y_train, dtype=torch.float32).view(-1, 1)
-y_test_tensor = torch.tensor(y_test, dtype=torch.float32).view(-1, 1)
-
-train_dataset = TensorDataset(X_train_tensor, y_train_tensor)
-test_dataset = TensorDataset(X_test_tensor, y_test_tensor)
-train_loader = DataLoader(train_dataset, batch_size=32, shuffle=True)
-test_loader = DataLoader(test_dataset, batch_size=32, shuffle=False)
-
-
-class IncomeClassifier(nn.Module):
-    def __init__(self):
-        super(IncomeClassifier, self).__init__()
-        self.layer1 = None
-        self.layer2 = nn.Linear(128, 64)
-        self.output = nn.Linear(64, 1)
-        self.relu = nn.ReLU()
-        self.sigmoid = nn.Sigmoid()
-        self.dropout = nn.Dropout(0.5)
-
-    def forward(self, x):
-        if self.layer1 is None:
-            self.initialize_model(x.size(1))  # Initialize model with input dimension
-        x = self.relu(self.layer1(x))
-        x = self.dropout(x)
-        x = self.relu(self.layer2(x))
-        x = self.dropout(x)
-        x = self.sigmoid(self.output(x))
-        return x
-
-    def initialize_model(self, input_dim):
-        self.layer1 = nn.Linear(input_dim, 128)
-
-
-model = IncomeClassifier()
-
-
-def train(model, train_loader, num_epochs=10):
-    criterion = nn.BCELoss()
-    optimizer = optim.Adam(model.parameters(), lr=0.001)
-    model.train()
-    for epoch in range(num_epochs):
-        for X_batch, y_batch in train_loader:
-            optimizer.zero_grad()
-            outputs = model(X_batch)
-            loss = criterion(outputs, y_batch)
-            loss.backward()
-            optimizer.step()
-        print(f"Epoch [{epoch+1}/{num_epochs}], Loss: {loss.item():.4f}")
-
-
-def evaluate(model, test_loader):
-    model.eval()
-    with torch.no_grad():
-        correct = 0
-        total = 0
-        for X_batch, y_batch in test_loader:
-            outputs = model(X_batch)
-            predicted = (outputs > 0.5).float()
-            total += y_batch.size(0)
-            correct += (predicted == y_batch).sum().item()
-        accuracy = correct / total
-        print(f"Accuracy: {accuracy:.2f}")
-
-
-train(model, train_loader, num_epochs=10)
-
-evaluate(model, test_loader)
